fix(pdt): fixes naming that usd means full usd value and this is what is being tracked.
this also rewrites tests to use 1e6 the amount, so the usd value is the same

+++ packages/ve3-tests/src/common/exec_asset_gauge.rs
@@ -75,7 +75,7 @@ impl TestingSuite {
   #[track_caller]
   pub fn e_gauge_add_rebase_in_ampluna(
     &mut self,
-    amount: u32,
+    amount: u128,
     result: impl Fn(Result<AppResponse, anyhow::Error>),
   ) -> &mut TestingSuite {
     let msg = ExecuteMsg::AddRebase {};
